// src/tmdb/schema.d.ts
export interface operations {
// ...
    "trending-movies": {
        parameters: {
            query?: {
                /** @description `ISO-639-1`-`ISO-3166-1` code */
                language?: string;
            };
            header?: never;
            path: {
                time_window: "day" | "week";
            };
            cookie?: never;
        };
        requestBody?: never;
        responses: {
            /** @description 200 */
            200: {
                headers: {
                    [name: string]: unknown;
                };
                content: {
                    "application/json": {
                        /**
                         * @default 0
                         * @example 1
                         */
                        page: number;
                        results?: {
                            /**
                             * @default true
                             * @example false
                             */
                            adult: boolean;
                            /** @example /44immBwzhDVyjn87b3x3l9mlhAD.jpg */
                            backdrop_path?: string;
                            /**
                             * @default 0
                             * @example 934433
                             */
                            id: number;
                            /** @example Scream VI */
                            title?: string;
                            /** @example en */
                            original_language?: string;
                            /** @example Scream VI */
                            original_title?: string;
                            /** @example Following the latest Ghostface killings, the four survivors leave Woodsboro behind and start a fresh chapter. */
                            overview?: string;
                            /** @example /wDWwtvkRRlgTiUr6TyLSMX8FCuZ.jpg */
                            poster_path?: string;
                            /** @example movie */
                            media_type?: "movie";
                            genre_ids?: number[];
                            /**
                             * @default 0
                             * @example 609.941
                             */
                            popularity: number;
                            /** @example 2023-03-08 */
                            release_date?: string;
                            /**
                             * @default true
                             * @example false
                             */
                            video: boolean;
                            /**
                             * @default 0
                             * @example 7.374
                             */
                            vote_average: number;
                            /**
                             * @default 0
                             * @example 684
                             */
                            vote_count: number;
                        }[];
                        /**
                         * @default 0
                         * @example 1000
                         */
                        total_pages: number;
                        /**
                         * @default 0
                         * @example 20000
                         */
                        total_results: number;
                    };
                };
            };
        };
    };
// ...
}